THF-405: Recreated Events search functionality.
